refactor directory layout and add more examples
