+++ sdk/storage/blob/2018-11-09/azblob/examples_test.go
