+++ sdk/storage/fileshare/2019-02-02/azfileshare/service.go
@@ -1,0 +1,1 @@
+package azfileshare
