+++ sdk/storage/mgmt/2019-04-01/azstorage/models.go
@@ -1,0 +1,1 @@
+package azstorage
