+++ sdk/storage/mgmt/2019-04-01/azstorage/service.go
@@ -1,0 +1,1 @@
+package azstorage
